rag_chatbot/requirements.txt
streamlit
langchain
langchain-community
chromadb
sentence-transformers
python-dotenv
openai
pypdf
docx2txt